Extract Node, remove receiver from Communication
