+++ node/src/main/scala/coop/rchain/node/conf.scala
@@ -1,0 +1,35 @@
+package coop.rchain.node
+
+import org.rogach.scallop._
+
+final case class Conf(arguments: Seq[String]) extends ScallopConf(arguments) {
+  version(s"RChain Node ${BuildInfo.version}")
+
+  val name =
+    opt[String](default = None, short = 'n', descr = "Node name or key.")
+
+  val port =
+    opt[Int](default = Some(30304), short = 'p', descr = "Network port to use.")
+
+  val httpPort =
+    opt[Int](default = Some(8080), short = 'x', descr = "HTTP port.")
+
+  val bootstrap =
+    opt[String](default = Some("rnode://0f365f1016a54747b384b386b8e85352@216.83.154.106:30012"),
+                short = 'b',
+                descr = "Bootstrap rnode address for initial seed.")
+
+  val standalone = opt[Boolean](default = Some(false),
+                                short = 's',
+                                descr = "Start a stand-alone node (no bootstrapping).")
+
+  val host = opt[String](default = None, descr = "Hostname or IP of this node.")
+
+  val repl = opt[Boolean](default = Some(false),
+                          short = 'r',
+                          descr = "Start node with REPL (but without P2P network)")
+  val eval = opt[String](default = None,
+                         descr = "Start node to evaluate rholang in file (but without P2P network)")
+
+  verify()
+}
